feat: 로그인 및 회원가입 기능 완성 및 관리자 기반 삭제기능 추가

+++ src/main/java/com/example/crawling/controller/StockController.java
@@ -3,11 +3,17 @@ package com.example.crawling.controller;
 import com.example.crawling.dto.CrawledPriceResponseDto;
 import com.example.crawling.dto.StockResponseDto;
 import com.example.crawling.model.CrawledPrice;
+import com.example.crawling.model.User;
+import com.example.crawling.repository.UserRepository;
 import com.example.crawling.service.StockService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Map;
 
@@ -17,10 +23,11 @@ import java.util.Map;
 public class StockController {
 
     private final StockService stockService;
+    private final UserRepository userRepository;
 
     @GetMapping
-    public Map<String, StockResponseDto> getStocks() {
-        return stockService.getStockData();
+    public Map<String, StockResponseDto> getStocks(Principal principal) {
+        return stockService.getStockData(principal);
     }
 
     // 키워드 검색 조회
@@ -38,8 +45,22 @@ public class StockController {
     }
 
     // 해당 코드의 모든 내용 삭제
+    @PreAuthorize("hasRole('ADMIN')")
     @DeleteMapping("/deleteByCode")
-    public ResponseEntity<Void> deleteStock(@RequestParam String stockCode) {
+    public ResponseEntity<?> deleteStock(@RequestParam String stockCode, Principal principal) {
+        String username = principal.getName();
+
+
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new RuntimeException("없는 사용자입니다."));
+
+        //관리자 권한 확인
+        if (!"ROLE_ADMIN".equals(user.getRole())) {
+            return ResponseEntity
+                    .status(HttpStatus.FORBIDDEN)
+                    .body("관리자만 삭제가능합니다.");
+        }
+
         stockService.deletePricesByStockCode(stockCode);
         return ResponseEntity.noContent().build();
     }
